feat: add Leetcode 39 solution
Combination Sum
Practice for backtracking (recursive). Time complexity consists of
having to choose to add the value or not (doubling), and worst case for
both complexities account for the number of times the smallest
candidate can fit into the target value (t/m). The space complexity
only accounts for the auxiliary memory since the output size, number of
valid solutions, is input dependent.