feat: add Leetcode 703 solution
Kth Largest Element in a Stream
Practice for heaps/priority queues. Implemented a min heap to keep
the kth largest value at the top of the heap. The resulting time and
space complexities are O(n log k) and O(k), respectively, where k is the
size of the heap.