feat: add Leetcode 133 solution
Clone Graph
Practice for graphs (connected undirected). Cloned graph using a DFS
helper function and referenced new/copied nodes with the old nodes in a
hashmap. Space complexity is O(v) extra for the hashmap and call stack,
but it is O(v+e) if you include the required output.