feat: add Leetcode 740 solution
Delete and Earn
Practice for DP. Implemented Bottom-Up DP similar to House Robber
problem. Time complexity is based on iterating through n values in nums
to create a counter and the for loop iterating from one to the max value
in nums. Space complexity is the number of unique values in the counter
(worst case is n).